Workaround for ld bug
Apparently, there is a bug in ld that if there is a gap in the address
space between .tdata and .tbss, for example due to alignment
requirements of .tbss, it will error out with the message "section
`.tbss' can't be allocated in segment 1".  Since .tbss will _probably_
not require more alignment than 16, pad out the .tdata section to 16
byte alignment to avoid the problem until it can be fixed properly.